scheduler/requirements.txt
httpx==0.27.2
psycopg2-binary==2.9.9
redis==5.0.8
yfinance==0.2.43
python-dotenv==1.0.1
apscheduler==3.10.4
pytz==2024.1
